Add RSVP form
Add an RSVP form which will send responses to a google form. Update CSS
to properly display the form and style the inputs and labels. Ad logic to
JavaScript to ensure that the correct data is sent to Google, and to
ensure that the form can not be submitted with missing requirements.